Deprecate BaseViewHolder and rename it to ItemViewHolder
